skia-org example: cleanup modules and separate drivers into drivers/
