+++ skia-safe/examples/skia-org/drivers/vulkan.rs
@@ -1,10 +1,68 @@
+use crate::artifact;
+use crate::drivers::DrawingDriver;
 use ash::version::{DeviceV1_0, EntryV1_0, InstanceV1_0};
 use ash::vk;
 use ash::vk::Handle;
 use ash::{Entry, Instance};
-use skia_safe::gpu;
+use skia_safe::{gpu, Budgeted, Canvas, ImageInfo, Surface};
 use std::ffi::{c_void, CString};
 use std::os::raw;
+use std::path::Path;
+use std::ptr;
+
+pub enum Vulkan {}
+
+impl DrawingDriver for Vulkan {
+    const NAME: &'static str = "vulkan";
+
+    fn draw_image(
+        (width, height): (i32, i32),
+        path: &Path,
+        name: &str,
+        func: impl Fn(&mut Canvas),
+    ) {
+        let ash_graphics = unsafe { AshGraphics::new("skia-org") };
+
+        let get_proc = |of| unsafe {
+            match ash_graphics.get_proc(of) {
+                Some(f) => f as _,
+                None => {
+                    println!("resolve of {} failed", of.name().to_str().unwrap());
+                    ptr::null()
+                }
+            }
+        };
+
+        let backend_context = unsafe {
+            gpu::vk::BackendContext::new(
+                ash_graphics.instance.handle().as_raw() as _,
+                ash_graphics.physical_device.as_raw() as _,
+                ash_graphics.device.handle().as_raw() as _,
+                (
+                    ash_graphics.queue_and_index.0.as_raw() as _,
+                    ash_graphics.queue_and_index.1,
+                ),
+                &get_proc,
+            )
+        };
+
+        let mut context = gpu::Context::new_vulkan(&backend_context).unwrap();
+
+        let image_info = ImageInfo::new_n32_premul((width * 2, height * 2), None);
+        let mut surface = Surface::new_render_target(
+            &mut context,
+            Budgeted::YES,
+            &image_info,
+            None,
+            gpu::SurfaceOrigin::TopLeft,
+            None,
+            false,
+        )
+        .unwrap();
+
+        artifact::draw_image_on_surface(&mut surface, path, name, func);
+    }
+}
 
 pub struct AshGraphics {
     pub entry: Entry,
