+++ skia-safe/examples/skia-org/main.rs
@@ -1,14 +1,8 @@
-use crate::artifact::DrawingDriver;
 use clap::{App, Arg};
 use offscreen_gl_context::{GLContext, GLVersion, NativeGLContext};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
-extern crate skia_safe;
-
-#[cfg(feature = "vulkan")]
-use crate::artifact::Vulkan;
-#[cfg(feature = "vulkan")]
-use crate::drivers::skia_ash::AshGraphics;
+use crate::drivers::DrawingDriver;
 
 #[cfg(feature = "vulkan")]
 #[macro_use]
@@ -16,211 +10,13 @@ extern crate ash;
 
 // TODO: think about making the examples more Rust-idiomatic, by using method chaining for Paint / Paths, for example.
 
+mod artifact;
 mod drivers;
 mod skcanvas_overview;
 mod skpaint_overview;
 mod skpath_overview;
 #[cfg(feature = "shaper")]
 mod skshaper_example;
-
-pub(crate) mod artifact {
-    use skia_safe::gpu;
-    use skia_safe::{Budgeted, Canvas, EncodedImageFormat, ImageInfo, Surface};
-    use std::fs;
-    use std::io::Write;
-    use std::path::Path;
-
-    #[cfg(feature = "vulkan")]
-    use crate::drivers::skia_ash::AshGraphics;
-    #[cfg(feature = "vulkan")]
-    use ash::version::InstanceV1_0;
-    #[cfg(feature = "vulkan")]
-    use ash::vk::Handle;
-    #[cfg(feature = "vulkan")]
-    use std::ptr;
-
-    pub trait DrawingDriver {
-        const NAME: &'static str;
-
-        fn draw_image<F>(size: (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> ();
-
-        fn draw_image_256<F>(path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            Self::draw_image((256, 256), path, name, func)
-        }
-    }
-
-    pub enum CPU {}
-    pub enum PDF {}
-    #[cfg(feature = "svg")]
-    pub enum SVG {}
-
-    pub enum OpenGL {}
-    #[cfg(feature = "vulkan")]
-    pub enum Vulkan {}
-
-    impl DrawingDriver for CPU {
-        const NAME: &'static str = "cpu";
-
-        fn draw_image<F>((width, height): (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            let mut surface = Surface::new_raster_n32_premul((width * 2, height * 2)).unwrap();
-            draw_image_on_surface(&mut surface, path, name, func);
-        }
-    }
-
-    impl DrawingDriver for PDF {
-        const NAME: &'static str = "pdf";
-
-        fn draw_image<F>(size: (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            let mut document = skia_safe::pdf::new_document(None).begin_page(size, None);
-            func(document.canvas());
-            let data = document.end_page().close();
-            write_file(data.as_bytes(), path, name, "pdf");
-        }
-    }
-
-    #[cfg(feature = "svg")]
-    impl DrawingDriver for SVG {
-        const NAME: &'static str = "svg";
-
-        fn draw_image<F>(size: (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            use skia_safe::Rect;
-            let mut canvas = skia_safe::svg::Canvas::new(Rect::from_size(size), None);
-            func(&mut canvas);
-            let data = canvas.end();
-            write_file(data.as_bytes(), path, name, "svg");
-        }
-    }
-
-    impl DrawingDriver for OpenGL {
-        const NAME: &'static str = "opengl";
-
-        fn draw_image<F>((width, height): (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            let mut context = gpu::Context::new_gl(None).unwrap();
-
-            let image_info = ImageInfo::new_n32_premul((width * 2, height * 2), None);
-            let mut surface = Surface::new_render_target(
-                &mut context,
-                Budgeted::YES,
-                &image_info,
-                None,
-                gpu::SurfaceOrigin::BottomLeft,
-                None,
-                false,
-            )
-            .unwrap();
-
-            draw_image_on_surface(&mut surface, path, name, func);
-        }
-    }
-
-    #[cfg(feature = "vulkan")]
-    impl DrawingDriver for Vulkan {
-        const NAME: &'static str = "vulkan";
-
-        fn draw_image<F>((width, height): (i32, i32), path: &Path, name: &str, func: F)
-        where
-            F: Fn(&mut Canvas) -> (),
-        {
-            let ash_graphics = unsafe { AshGraphics::new("skia-org") };
-
-            let get_proc = |of| unsafe {
-                match ash_graphics.get_proc(of) {
-                    Some(f) => f as _,
-                    None => {
-                        println!("resolve of {} failed", of.name().to_str().unwrap());
-                        ptr::null()
-                    }
-                }
-            };
-
-            let backend_context = unsafe {
-                gpu::vk::BackendContext::new(
-                    ash_graphics.instance.handle().as_raw() as _,
-                    ash_graphics.physical_device.as_raw() as _,
-                    ash_graphics.device.handle().as_raw() as _,
-                    (
-                        ash_graphics.queue_and_index.0.as_raw() as _,
-                        ash_graphics.queue_and_index.1,
-                    ),
-                    &get_proc,
-                )
-            };
-
-            let mut context = gpu::Context::new_vulkan(&backend_context).unwrap();
-
-            let image_info = ImageInfo::new_n32_premul((width * 2, height * 2), None);
-            let mut surface = Surface::new_render_target(
-                &mut context,
-                Budgeted::YES,
-                &image_info,
-                None,
-                gpu::SurfaceOrigin::TopLeft,
-                None,
-                false,
-            )
-            .unwrap();
-
-            draw_image_on_surface(&mut surface, path, name, func);
-        }
-    }
-
-    fn draw_image_on_surface<F>(surface: &mut Surface, path: &Path, name: &str, func: F)
-    where
-        F: Fn(&mut Canvas) -> (),
-    {
-        let mut canvas = surface.canvas();
-
-        canvas.scale((2.0, 2.0));
-        func(&mut canvas);
-        let image = surface.image_snapshot();
-        let data = image.encode_to_data(EncodedImageFormat::PNG).unwrap();
-        write_file(data.as_bytes(), path, name, "png");
-    }
-
-    fn write_file(bytes: &[u8], path: &Path, name: &str, ext: &str) {
-        fs::create_dir_all(&path).expect("failed to create directory");
-
-        let mut file_path = path.join(name);
-        file_path.set_extension(ext);
-
-        let mut file = fs::File::create(file_path).expect("failed to create file");
-        file.write_all(bytes).expect("failed to write to file");
-    }
-}
-
-pub(crate) mod resources {
-
-    use skia_safe::{Data, Image};
-
-    pub fn color_wheel() -> Image {
-        let bytes = include_bytes!("resources/color_wheel.png");
-        let data = Data::new_copy(bytes);
-        Image::from_encoded(data, None).unwrap()
-    }
-
-    pub fn mandrill() -> Image {
-        let bytes = include_bytes!("resources/mandrill_512.png");
-        let data = Data::new_copy(bytes);
-        Image::from_encoded(data, None).unwrap()
-    }
-}
 
 fn main() {
     const OUT_PATH: &str = "OUT_PATH";
@@ -238,7 +34,7 @@ fn main() {
             Arg::with_name(DRIVER)
                 .long(DRIVER)
                 .takes_value(true)
-                .possible_values(get_possible_drivers().as_slice())
+                .possible_values(get_available_drivers().as_slice())
                 .multiple(true)
                 .help("In addition to the CPU, render with the given driver."),
         )
@@ -258,22 +54,22 @@ fn main() {
         }
     };
 
-    if drivers.contains(&artifact::CPU::NAME) {
-        draw_all::<artifact::CPU>(&out_path);
+    if drivers.contains(&drivers::CPU::NAME) {
+        draw_all::<drivers::CPU>(&out_path);
     }
 
-    if drivers.contains(&artifact::PDF::NAME) {
-        draw_all::<artifact::PDF>(&out_path);
+    if drivers.contains(&drivers::PDF::NAME) {
+        draw_all::<drivers::PDF>(&out_path);
     }
 
     #[cfg(feature = "svg")]
     {
-        if drivers.contains(&artifact::SVG::NAME) {
-            draw_all::<artifact::SVG>(&out_path);
+        if drivers.contains(&drivers::SVG::NAME) {
+            draw_all::<drivers::SVG>(&out_path);
         }
     }
 
-    if drivers.contains(&artifact::OpenGL::NAME) {
+    if drivers.contains(&drivers::OpenGL::NAME) {
         let context = GLContext::<NativeGLContext>::create(
             sparkle::gl::GlType::Gl,
             GLVersion::MajorMinor(3, 3),
@@ -282,11 +78,14 @@ fn main() {
         .unwrap();
 
         context.make_current().unwrap();
-        draw_all::<artifact::OpenGL>(&out_path);
+        draw_all::<drivers::OpenGL>(&out_path);
     }
 
     #[cfg(feature = "vulkan")]
     {
+        use drivers::vulkan::AshGraphics;
+        use drivers::Vulkan;
+
         if drivers.contains(&Vulkan::NAME) {
             match AshGraphics::vulkan_version() {
                 Some((major, minor, patch)) => {
@@ -295,11 +94,11 @@ fn main() {
                 None => println!("Failed to detect Vulkan version, falling back to 1.0.0"),
             }
 
-            draw_all::<artifact::Vulkan>(&out_path)
+            draw_all::<Vulkan>(&out_path)
         }
     }
 
-    fn draw_all<Driver: DrawingDriver>(out_path: &PathBuf) {
+    fn draw_all<Driver: DrawingDriver>(out_path: &Path) {
         let out_path = out_path.join(Driver::NAME);
 
         skcanvas_overview::draw::<Driver>(&out_path);
@@ -311,7 +110,7 @@ fn main() {
     }
 }
 
-fn get_possible_drivers() -> Vec<&'static str> {
+fn get_available_drivers() -> Vec<&'static str> {
     let mut drivers = vec!["cpu", "pdf", "opengl"];
     if cfg!(feature = "vulkan") {
         drivers.push("vulkan")
@@ -320,4 +119,21 @@ fn get_possible_drivers() -> Vec<&'static str> {
         drivers.push("svg");
     }
     drivers
+}
+
+pub(crate) mod resources {
+
+    use skia_safe::{Data, Image};
+
+    pub fn color_wheel() -> Image {
+        let bytes = include_bytes!("resources/color_wheel.png");
+        let data = Data::new_copy(bytes);
+        Image::from_encoded(data, None).unwrap()
+    }
+
+    pub fn mandrill() -> Image {
+        let bytes = include_bytes!("resources/mandrill_512.png");
+        let data = Data::new_copy(bytes);
+        Image::from_encoded(data, None).unwrap()
+    }
 }
